requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cycler==0.12.1
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.2
h11==0.14.0
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
onnxruntime==1.21.0
opencv-python==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pefile==2023.2.7
pillow==11.1.0
protobuf==6.30.2
psutil==7.0.0
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pygame==2.6.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.2
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
selenium==4.30.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
starlette==0.46.1
sympy==1.13.1
thop==0.1.1.post2209072238
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
uvicorn==0.34.0
websocket-client==1.8.0
wsproto==1.2.0
